chore(rig-candle): move inline test modules to sibling files
